* Source compiles under Linux ** Add Makefile
