--- RTC.H
@@ -1,9 +1,0 @@
-void rtc (FILE *screen)
-{
-   fflush (stdin);
-   fprintf (screen, "\nHit <RETURN> to continue: ");
-   getchar ();
-   fflush (stdin);
-   fprintf (screen, "\n");
-}
-
