+++ src/rtc.h
@@ -1,0 +1,9 @@
+void rtc (FILE *screen)
+{
+   fflush (stdin);
+   fprintf (screen, "\nHit <RETURN> to continue: ");
+   getchar ();
+   fflush (stdin);
+   fprintf (screen, "\n");
+}
+
